exp/services/webauth: remove the stellar.toml domain checks (#3113)
### What
Remove the stellar.toml domain checks that occur on application start. Essentially revert the changes relating to this specific feature that were introduced in 5e82f16273b7ce97e50866cba207bc52f498e9c7.

### Why
SEP-10 v2.0.0 highlighted an existing feature that the SIGNING_KEY field in the stellar.toml be the key that the SEP-10 server uses for signing. In 5e82f16273b7ce97e50866cba207bc52f498e9c7 we added the feature that the server would verify that the stellar.toml that references the SEP-10 server. This feature is inconvenient because not all SEP-10 services have a stellar.toml. It's also scope creep on what this SEP-10 server implementation is concerned with.