clients/horizonclient,txnbuild: update changelogs for v4.1.0 (#3136)
Update the changelogs for clients/horizonclient and txnbuild for the combined v4.1.0 release. Also add a missing empty release details to clients/horizonclient changelog that was missing in the previous release which only changed txnbuild.

There are several changes to txnbuild on master that could be released. I specifically want to get the SEP-10 v2.1.0 changes in a release.

There are only meaningful changes to txnbuild but txnbuild and horizonclient get released together, so that is why there are empty changelog release notes for the latter.