ingest/ledgerbackend: Remove returning error on Stellar-Core process exit during catchup (#3260)
Fixes a bug introduced in a10c000 because of which `PrepareRange` and
`GetLedger` methods could return an error after Stellar-Core process
exit but before all ledgers are read from the buffer. To fix it we now
handle process exit in `bufferedLedgerMetaReader` only and only in case
of errors. In `PrepareRange` we return error only on Stellar-Core exit
with error. This won't work with user initiated shutdown, it will be
fixed in #3258.