all: enable codeql code analysis (#3390)
### What

Enable GitHub's CodeQL static analysis to run periodically and on PRs to `master`. It will open issues in the Security tab when it finds problems.

### Why

To get feedback about issues we might miss.

I wanted to enable this a while ago but there was no Go support, but I just stumbled on an advertisement in the GitHub UI suggesting it for this repo. I'm not sure how long ago they added support, but here we are. The yml file was mostly created by the GitHub UI suggestion box, but I edited it to include `release-*` branches.