txnbuild: impl SEP-10 v3.1 add web_auth_domain support (#3359)
### What

Implement SEP-10 v3.1, adding support for building and verifying SEP-10 challenge transactions with the web_auth_domain manage data operation.

### Why

To maintain compatibility with SEP-10.

Close #3334.

### Compatibility

This SEP-10 change is completely backwards compatible with any server or client configuration using SEP-10 v2.1 or greater. If a server is on v3.1 and a client is on v2.1, they are compatible. If a server is on v2.1 and a client is on v3.1, they are compatible.

The SDK changes are however not backwards compatible because of changes to exported function signatures in the `txnbuild` package. Because of this, this change needs releasing in a major release of txnbuild.

The webauth changes are not backwards compatible because the server requires the `--domain` parameter to be specified for successful startup. This does not affect how webauth gets released because it is experimental. This will require updates to any deployments using webauth.