Remove monorepo Dockerfile (#3425)
Each individual service in the monorepo has their own Dockerfile (e.g. services/horizon/docker/Dockerfile). There are no CircleCI jobs or scripts which rely on the root Dockerfile that I'm aware of. It seems to me that this Dockerfile is effectively dead code and that removing it would reduce potential confusion.