historyarchive: Archivist slow scan missing (#3528)
This change fixes a bug in archivist that causes it to mis-report archives as
"not missing files" when they are, in fact, missing files. Fairly serious
condition. It's caused on archives accessed through http -- which resorts to a
slow-scan path that probes each file one by one using HEAD -- rather than
archives like s3 that can be fast-scanned with a list command.

Existing behaviour is wrong. It treats "we got information about a file at all"
as meaning "the file is there", even when the information we got is "the file is
not there". Classic sort of 3-value-logic mistake on my part.