services/horizon: Removed LedgerBackend calls inside ingest critical section (#3518)
This commit moves calls to `LedgerBackend` in `buildState` and `resumeState`
outside critical section which is a DB transaction acquiring ingestion lock.

To clean the code `LedgerBackend.GetLedgerBlocking` method was introduced that
always blocks until the ledger is available in the backend. Additionally,
`GetLatestLedger` method name was changed to `GetLatestHistoryLedger` for
clarity and `ProcessorRunner` methods have changed and do not depend on
`LedgerBackend` anymore.

This is done to prevent situations in which `LedgerBackend` hangs (ex. broken
connection to Core DB or remote captive core) inside a critical section
ingestion will halt in entire cluster. With this change only a single ingesting
node will halt.