txnbuild: fix bug where base fee is not allowed to be zero (#3622)
Allow the base fee to be zero when building transactions.

The `txnbuild.NewTransaction` function errors when creating a transaction with a base fee of zero.

This is a bug.

Transactions with a zero base fee are accepted to the Stellar network when they are wrapped in a fee bump transaction envelope that pays the fee of the transaction. Applications use a zero base fee on the inner transaction to support fee sponsorship by an account that isn't the transaction source account.