services/regulated-assets-approval-server: refactor tests targeting the /tx-approve endpoint (#3647)
### What

Refactor tests addressing the `/tx-approve` endpoint for revised/rejected/action_required responses.

Now the unit tests are covering not only the happy path but also all edge cases from the transaction approval logic. On the other hand, the API tests were reduced to cover the happy path only, so they become easier to maintain and we don't repeat ourselves (DRY).

Additionally, removed unnecessary setup in the tests where variables and mocked functions were being declared without need.

### Why

There was confusion between unit tests and API tests in this service so the package was not properly tested in terms of unit tests and a lot of the edge cases were being validated only in the API tests.

This makes the code and tests harder to maintain as a developer would need to maintain package-related tests outside the package.