services/horizon: Return 503 error in POST /transactions if Stellar-Core not synced (#3653)
This commit changes `POST /transactions` behaviour to return `503 Service
Unavailable` (`stale_history`) instead of `504 Gateway Timeout` (`timeout`)
response if connected Stellar-Core is not synced.

`timeout` error is confusing when the connected Stellar-Core is not synced
because it means that the transaction has been broadcasted but there is no
result yet. What actually happens is that the transaction is not broadcasted
(and Horizon cannot ingest it) when Stellar-Core is not in synced. The new
response is returned as soon as Horizon determines that the Stellar-Core status
has changed.