services/horizon/txsub: Batch status check queries for pending transactions in txsub (#3563)
This commit modifies `txsub.Tick` to check the status of transactions in the
queue by sending a batch query instead of sending separate queries to get status
of each transaction in the queue.

This was done due performance reasons. If the number of transactions in the
queue is large this slows down entire `Tick` function. Data in #3552 suggest
that previous method can take even 4s per `Tick` execution for around 1000
transactions in the queue.