all: reduce go.list to deps used during build (#3768)
### What
Shorten the list of dependencies in `go.list` to be dependencies that are used during the build of code and tests.

### Why
Our `go.list` file contains all dependencies in the graph, even dependencies that don't make it to code. It makes assessing the actual impact of upgrades like #3767 difficult.

### Known Limitations
Moving away from using the `-m` flag means tags apply. Any code behind a build tag won't be included. It's not trivial to use `go list` with multiple tags. There's an issue open at golang/go#42504 that will hopefully make it possible to run the go list command with all tags included.