keypair: remove interface comparisons (#3782)
Limit the keypair types to be comparable to themselves, so that a FromAddress is only comparable to a FromAddress and a Full is only comparable to a Full.

The functions don't actually provide this functionality even though they accept a KP interface type. In the case different types are passed in they return false, so it isn't useful that they accept the interface. The fact they accept the interface makes their use error prone. It's really easy to pass in a Full thinking it's a FromAddress, and it will compile but never return true. I ran into this problem myself and it took me a while to debug in a small sized application. Given that it never returns true we should just never support passing in the other types.

There could be some value in retaining that all KPs have an Equal that works with any KP, but I think the value delivered with that is small compared to the foot gun problem described above.

This is a breaking change but it is a breaking change to functionality that only exists on master and is unreleased, so no compatibility promise implicit or explicit exists.